pause AI in opponent's turn (Temporary)
This is a temporary and not entirely effective solution, to prevent AI from spamming CPU useage to make their opponents time out.
Yes I could use renice instead, but its just as easy for an AI to get around.

I will have to replace this with a better solution. But I don't have one at the moment.