[PATCH] Allow Human_Controller to move scouts further
Thanks to David Gow (sulix) for this patch.

I was aware of the limitation, but decided to wait until someone complained before fixing it (I also overestimated the effort required to fix it).
So I'd like to congratulate David for being the first person to actually try and use the manager program!

I'll have to allocate a prize for that.