Fixed bug in Java AI
In KillUnit:
I was using Vector.remove(Object), but I realised that this doesn't compare
references, instead it uses "compareTo". Which meant the AI deleted every unit
as soon as the first one died. Fun.

I discovered gcj, which lets you compile a Java program.
This is much easier than trying to convince the java interpreter to run.